fix -Wshorten-64-to-32 warning
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@38172 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/tape/t64.c
@@ -449,7 +449,7 @@ int t64_read(t64_t *t64, uint8_t *buf, size_t size)
     }
     t64->current_file_seek_position += (int)amount;
 
-    return amount;
+    return (int)amount;
 }
